requirements.txt
python-dotenv
openai
anthropic
google-genai
datasets
transformers
torch
tqdm
